161026 Naver Map

+++ app/src/main/java/com/daejong/seoulpharm/util/NetworkManager.java
@@ -91,7 +91,7 @@ public class NetworkManager {
     // 좌표 > 주소 변환
     public static final String CONVERT_TO_ADDRESS_URL = "https://openapi.naver.com/v1/map/reversegeocode";
 
-    public void getAddress(Context context, long latitude, long longtitude, final OnResultListener<String> listener) {
+    public void getAddress(Context context, double latitude, double longtitude, final OnResultListener<String> listener) {
         RequestParams params = new RequestParams();
         params.put("query", longtitude + "," + latitude);
 
@@ -112,6 +112,8 @@ public class NetworkManager {
             }
         });
     }
+
+
 
     // naver transrator
     public static final String TRANSLATION_URL = "https://openapi.naver.com/v1/language/translate";
